+++ enchanter/addons/activation.pyi
