+++ tests/addons/test_activations.py
@@ -1,11 +1,11 @@
 import torch
 import torch.jit
 import numpy as np
-import enchanter.addon as addon
+import enchanter.addons as addons
 
 
 def test_swish1():
-    swish = torch.jit.script(addon.Swish())
+    swish = torch.jit.script(addons.Swish())
     x = torch.tensor([1.0], dtype=torch.float32)
     out = swish(x)
     out = np.round(out.detach().numpy(), 4)
@@ -13,7 +13,7 @@ def test_swish1():
 
 
 def test_swish2():
-    swish = torch.jit.script(addon.Swish(beta=True))
+    swish = torch.jit.script(addons.Swish(beta=True))
     x = torch.tensor([1.0], dtype=torch.float32)
     out = swish(x)
     out = np.round(out.detach().numpy(), 4)
@@ -21,7 +21,7 @@ def test_swish2():
 
 
 def test_mish1():
-    mish = torch.jit.script(addon.Mish())
+    mish = torch.jit.script(addons.Mish())
     x = torch.tensor([1.0], dtype=torch.float32)
     out = mish(x)
     out = np.round(out.detach().numpy(), 4)
@@ -30,6 +30,6 @@ def test_mish1():
 
 def test_mish2():
     x = torch.tensor([1.0], dtype=torch.float32)
-    out = addon.mish(x)
+    out = addons.mish(x)
     out = np.round(out.detach().numpy(), 4)
     assert out == np.array([0.8651]).astype(np.float32)
